Convert character height to inches
For consistency, storing them in metric is impractical.

+++ src/com/lilithsthrone/game/character/GameCharacter.java
@@ -14471,10 +14471,7 @@ public abstract class GameCharacter implements XMLSaving {
 		return BodyShape.valueOf(Muscle.valueOf(getMuscleValue()), BodySize.valueOf(getBodySizeValue()));
 	}
 
-	/**
-	 * @return The character's height in cm.
-	 */
-	public int getHeightValue() {
+	public float getHeightValue() {
 		return body.getHeightValue();
 	}
 	
@@ -14486,18 +14483,18 @@ public abstract class GameCharacter implements XMLSaving {
 		return this.getHeightValue()<Height.getShortStatureCutOff();
 	}
 	
-	public int getMinimumHeight() {
+	public float getMinimumHeight() {
 		return this.getSubspecies().isShortStature()?Height.NEGATIVE_TWO_MIMIMUM.getMinimumValue():Height.ZERO_TINY.getMinimumValue();
 	}
 	
-	public int getMaximumHeight() {
+	public float getMaximumHeight() {
 		return Height.SEVEN_COLOSSAL.getMaximumValue();
 	}
 	
 	/**
 	 * @return Formatted description of height change.
 	 */
-	public String setHeight(int height) {
+	public String setHeight(float height) {
 		height = Math.min(Height.SEVEN_COLOSSAL.getMaximumValue(), Math.max(this.getSubspecies().isShortStature()?Height.NEGATIVE_TWO_MIMIMUM.getMinimumValue():Height.ZERO_TINY.getMinimumValue(), height));
 		
 		if (body.getHeightValue() < height) {
@@ -14543,7 +14540,7 @@ public abstract class GameCharacter implements XMLSaving {
 	/**
 	 * @return Formatted description of height change.
 	 */
-	public String incrementHeight(int increment) {
+	public String incrementHeight(float increment) {
 		return setHeight(getHeightValue() + increment);
 	}
